Create serialization.java
Write a java program that demonstrates object serialization and deserialization. Create a simple class named person with attributes name and age. Serialize an object of this class and write it to a file named “person.ser”. Then, read the serialized object from the file, deserialize it, and print the details of the deserialized object.